% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtract.R
\name{as.data.frame.extract}
\alias{as.data.frame.extract}
\title{Coerce extracted heritabilities to data.frame}
\usage{
\method{as.data.frame}{extract}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{An object of class extract.}

\item{row.names}{An optional character vector of row.names to be added to
the data.frame.}

\item{optional}{Ignored.}

\item{...}{Ignored.}
}
\description{
Function for coercing heritabilities extracted from an SSA object to a
data.frame for nicer printing and easier post processing.
}
