% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlot.R
\name{boxPlot}
\alias{boxPlot}
\title{Helper function for creating a map plot}
\usage{
boxPlot(x, trials, traits, title, output, ...)
}
\description{
Helper function for creating a map plot for an object of class TD.
}
\keyword{internal}
