% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{corPlot}
\alias{corPlot}
\title{Helper function for creating a correlation plot}
\usage{
corPlot(x, trials, traits, title, output, ...)
}
\description{
Helper function for creating a correlation plot for an object of class TD.
}
\keyword{internal}
