% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtract.R
\name{extractS3}
\alias{extractS3}
\alias{createExtract}
\alias{as.data.frame.extract}
\title{S3 class Extract}
\usage{
createExtract(result, what)

\method{as.data.frame}{extract}(x, row.names = NULL, optional = FALSE,
  ...)
}
\description{
Function for creating objects of S3 class Extract.
}
\author{
Bart-Jan van Rossum
}
\concept{extract}
\keyword{internal}
