% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSTA.R
\name{extractSTA}
\alias{extractSTA}
\title{Extract statistics from fitted models}
\usage{
extractSTA(
  STA,
  trials = names(STA),
  traits = NULL,
  what = "all",
  asDataFrame = length(what) == 1 && what != "all",
  keep = NULL,
  restoreColNames = FALSE
)
}
\arguments{
\item{STA}{An object of class STA.}

\item{trials}{A character vector of trials for which the statistics should be
computed. If not supplied, statistics are computed for all trials that have
been modeled.}

\item{traits}{A character vector of traits for which the statistics should be
computed. If not supplied, statistics are computed for all traits that have
been modeled.}

\item{what}{A character vector indicating which statistics should be
computed. Most statistics are available for all models, some only for models
fitted using a certain engine. If this is the case, this is indicated in the
list with options in details.\cr
If \code{what = "all"}, all available statistics are computed.}

\item{asDataFrame}{Should the output be reshaped to a data.frame. This is
only possible if the number of statistics to extract is one.}

\item{keep}{A character vector of column(s) in the object of class
\code{\link{TD}} used for modeling. These columns will be kept as output when
computing fitted values, residuals, standardized residuals and rMeans.
Columns can also be kept when computing (se)BLUEs and (se)BLUPs but only if
the column to keep contains unique values for the modeled variables, i.e. a
column repId with several different values per genotype cannot be kept.}

\item{restoreColNames}{Should the original column names be restored in the
output of the extracted data?}
}
\value{
A list with, per trial for which statistics have been extracted, a
list of those statistics.
}
\description{
Extract and calculate various results for fitted models such as BLUEs, BLUPs,
unit errors and heritabilities. For a full list of results that can be
extracted, see Details.\cr\cr
The result(s) to extract are specified in \code{what}. If a single result is
extracted, if possible, this result is returned as a data.frame. If this is
not possible, the result is returned as a list. If multiple results are
extracted at the same time, these are always returned as a list. Where
relevant, this behavior can be changed by specifying \code{asDataFrame}.
Results that are returned as data.frame are marked as such in the data.frame
column in the table in Details.\cr\cr
Most results can only be calculated if a model is fitted with genotype as
fixed or with genotype as random. E.g. to compute heritabilies a model should
be fitted with genotype as random effect. This is indicated in the list
in Details with "F" and "R" respectively.
}
\details{
Possible options for \code{what} are:
}
\examples{
## Fit model using SpATS.
myModel <- fitTD(TD = TDHeat05, design = "res.rowcol", traits = "yield")

## Extract all available statistics from the fitted model.
extr <- extractSTA(myModel)

## Extract only the BLUEs from the fitted model.
BLUEs <- extractSTA(myModel, what = "BLUEs")

## Extract only the BLUEs from the fitted model and keep trial as variable in
## the output.
BLUEs2 <- extractSTA(myModel, what = "BLUEs", keep = "trial")

}
\seealso{
\code{\link{fitTD}}
}
