% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutPlot.R
\name{layoutPlot}
\alias{layoutPlot}
\title{Helper function for creating a layout plot}
\usage{
layoutPlot(x, trials, traits, title, output, ...)
}
\description{
Helper function for creating a layout plot for an object of class TD.
}
\keyword{internal}
