% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlot.R
\name{mapPlot}
\alias{mapPlot}
\title{Helper function for creating a map plot}
\usage{
mapPlot(x, title = "Trial locations", output = TRUE, ...)
}
\description{
Helper function for creating a map plot for an object of class TD.
}
\keyword{internal}
