% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Helper function for creating a scatter plot}
\usage{
scatterPlot(x, trials, traits, title, output, ...)
}
\description{
Helper function for creating a scatter plot for an object of class TD.
}
\keyword{internal}
