% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenSSA.R
\docType{package}
\name{statgenSSA-package}
\alias{statgenSSA}
\alias{statgenSSA-package}
\title{statgenSSA: Single Site Analysis (SSA) of Field Trials}
\description{
The statgen packages, commissioned by the Generation Challenge 
    Program (GCP), are being developed to streamline the analysis of typical 
    plant breeding experiments, including both phenotypic analysis of field 
    trials and genotype-by-environment analysis. statgenSSA can be used for 
    phenotypic analysis of field trials using spatial models.
    Some functions have been created to be used in conjunction with the R 
    package 'asreml' for the 'ASReml' software, which can be obtained upon 
    purchase from 'VSN' international (<http://www.vsni.co.uk/software/asreml-r>).
}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl}

Other contributors:
\itemize{
  \item Fred van Eeuwijk (\href{https://orcid.org/0000-0003-3672-2921}{ORCID}) [contributor]
  \item Martin Boer [contributor]
  \item Daniela Bustos-Korts (\href{https://orcid.org/0000-0003-3827-6726}{ORCID}) [contributor]
  \item Emilie Millet (\href{https://orcid.org/0000-0002-2913-4892}{ORCID}) [contributor]
}

}
\keyword{internal}
