% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenSTA.R
\docType{package}
\name{statgenSTA-package}
\alias{statgenSTA}
\alias{statgenSTA-package}
\title{statgenSTA: Single Trial Analysis (STA) of Field Trials}
\description{
Phenotypic analysis of field trials using mixed models with and without spatial components. One of a series of statistical genetic packages for streamlining the analysis of typical plant breeding experiments developed by Biometris. Some functions have been created to be used in conjunction with the R package 'asreml' for the 'ASReml' software, which can be obtained upon purchase from 'VSN' international (\url{https://vsni.co.uk/software/asreml-r}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometris.github.io/statgenSTA/index.html}
  \item \url{https://github.com/Biometris/statgenSTA/}
  \item Report bugs at \url{https://github.com/Biometris/statgenSTA/issues}
}

}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Other contributors:
\itemize{
  \item Fred van Eeuwijk (\href{https://orcid.org/0000-0003-3672-2921}{ORCID}) [contributor]
  \item Martin Boer (\href{https://orcid.org/0000-0002-1879-4588}{ORCID}) [contributor]
  \item Marcos Malosetti (\href{https://orcid.org/0000-0002-8150-1397}{ORCID}) [contributor]
  \item Daniela Bustos-Korts (\href{https://orcid.org/0000-0003-3827-6726}{ORCID}) [contributor]
  \item Emilie J. Millet (\href{https://orcid.org/0000-0002-2913-4892}{ORCID}) [contributor]
  \item Joao Paulo (\href{https://orcid.org/0000-0002-4180-0763}{ORCID}) [contributor]
  \item Maikel Verouden (\href{https://orcid.org/0000-0002-4893-3323}{ORCID}) [contributor]
  \item Willem Kruijer (\href{https://orcid.org/0000-0001-7179-1733}{ORCID}) [contributor]
  \item Ron Wehrens (\href{https://orcid.org/0000-0002-8798-5599}{ORCID}) [contributor]
  \item Choazhi Zheng (\href{https://orcid.org/0000-0001-6030-3933}{ORCID}) [contributor]
}

}
\keyword{internal}
